Ticket #2869 Refactor experiment page - Generalization: GeneTooltipProperty is a separate class NameValuePair<T>... now
