+++ atlas-web/src/main/java/uk/ac/ebi/gxa/web/ui/NameValuePair.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2008-2011 Microarray Informatics Team, EMBL-European Bioinformatics Institute
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *
+ * For further details of the Gene Expression Atlas project, including source code,
+ * downloads and documentation, please see:
+ *
+ * http://gxa.github.com/gxa
+ */
+
+package uk.ac.ebi.gxa.web.ui;
+
+import org.codehaus.jackson.annotate.JsonProperty;
+
+/**
+ * @author Olga Melnichuk
+ */
+public class NameValuePair<V> {
+    private final String name;
+    private final V value;
+
+    public NameValuePair(String name, V value) {
+        this.name = name;
+        this.value = value;
+    }
+
+    @JsonProperty("name")
+    public String getName() {
+        return name;
+    }
+
+    @JsonProperty("value")
+    public V getValue() {
+        return value;
+    }
+}
